Разделить reducers на несколько файлов
