+++ src/reducer/game/game.js
@@ -1,8 +1,6 @@
 const initialState = {
   mistakes: 0,
-  step: -1,
-  questions: [],
-  isAuthorizationRequired: false
+  step: -1
 };
 
 const BusinessLogic = {
@@ -17,20 +15,9 @@ const BusinessLogic = {
   }
 };
 
-const Operation = {
-  loadQuestions: () => (dispatch, _getState, api) => {
-    return api.get(`/questions`)
-      .then((response) => {
-        dispatch(ActionCreators.loadQuestions(response.data));
-      });
-  }
-};
-
 const ActionTypes = {
   INCREMENT_MISTAKE: `INCREMENT_MISTAKE`,
   INCREMENT_STEP: `INCREMENT_STEP`,
-  LOAD_QUESTIONS: `LOAD_QUESTIONS`,
-  REQUIRED_AUTHORIZATION: `REQUIRED_AUTHORIZATION`,
   RESET_STATE: `RESET_STATE`
 };
 
@@ -60,20 +47,6 @@ const ActionCreators = {
     };
   },
 
-  loadQuestions: (questions) => {
-    return {
-      type: ActionTypes.LOAD_QUESTIONS,
-      payload: questions
-    };
-  },
-
-  requiredAuthorization: (status) => {
-    return {
-      type: ActionTypes.REQUIRED_AUTHORIZATION,
-      payload: status
-    };
-  },
-
   resetState: () => {
     return {
       type: ActionTypes.RESET_STATE
@@ -97,18 +70,6 @@ const reducer = (state = initialState, action) => {
       });
       break;
 
-    case ActionTypes.LOAD_QUESTIONS:
-      Object.assign(updatedState, state, {
-        questions: action.payload
-      });
-      break;
-
-    case ActionTypes.REQUIRED_AUTHORIZATION:
-      Object.assign(updatedState, state, {
-        isAuthorizationRequired: action.payload
-      });
-      break;
-
     case ActionTypes.RESET_STATE:
       Object.assign(updatedState, initialState);
       break;
@@ -122,7 +83,6 @@ const reducer = (state = initialState, action) => {
 };
 
 export {
-  Operation,
   ActionTypes,
   ActionCreators,
   BusinessLogic,
