+++ src/reducer/game/game.test.js
@@ -1,11 +1,7 @@
-import MockAdapter from 'axios-mock-adapter';
-import {createAPI} from '../api';
 import {
-  // ActionCreators,
   ActionTypes,
-  Operation,
   reducer
-} from './reducer';
+} from './game';
 
 describe(`Reducer works correctly`, () => {
   it(`Should return initial state by default`, () => {
@@ -65,25 +61,5 @@ describe(`Reducer works correctly`, () => {
       questions: [],
       isAuthorizationRequired: false
     });
-  });
-
-  it(`Should make a correct API call to /questions`, () => {
-    const dispatch = jest.fn();
-    const api = createAPI(dispatch);
-    const apiMock = new MockAdapter(api);
-    const questionLoader = Operation.loadQuestions();
-
-    apiMock
-      .onGet(`/questions`)
-      .reply(200, [{fake: true}]);
-
-    return questionLoader(dispatch, jest.fn(), api)
-      .then(() => {
-        expect(dispatch).toHaveBeenCalledTimes(1);
-        expect(dispatch).toHaveBeenNthCalledWith(1, {
-          type: ActionTypes.LOAD_QUESTIONS,
-          payload: [{fake: true}]
-        });
-      });
   });
 });
